+++ src/lib_clic/tezos_clic.ml
@@ -916,6 +916,40 @@ let args24 a b c d e f g h i j k l m n o p q r s t u v w x =
         (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x))
     (args2 (args16 a b c d e f g h i j k l m n o p) (args8 q r s t u v w x))
 
+let args25 a b c d e f g h i j k l m n o p q r s t u v w x y =
+  map_arg
+    ~f:
+      (fun _
+           ( (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p),
+             (q, r, s, t, u, v, w, x, y) ) ->
+      Lwt_result_syntax.return
+        ( a,
+          b,
+          c,
+          d,
+          e,
+          f,
+          g,
+          h,
+          i,
+          j,
+          k,
+          l,
+          m,
+          n,
+          o,
+          p,
+          q,
+          r,
+          s,
+          t,
+          u,
+          v,
+          w,
+          x,
+          y ))
+    (args2 (args16 a b c d e f g h i j k l m n o p) (args9 q r s t u v w x y))
+
 let switch ~doc ?short ~long () = Switch {doc; label = {long; short}}
 
 (* Argument parsing *)