+++ src/proto_018_Proxford/lib_sc_rollup_node/fueled_pvm.ml
@@ -34,7 +34,8 @@ module Make_fueled (F : Fuel.S) : FUELED_PVM with type fuel = F.t = struct
 
   type pvm_state = Irmin_context.tree
 
-  let get_reveal ~dac_client ~data_dir ~pvm_kind reveal_map hash =
+  let get_reveal ~dac_client ~pre_images_endpoint ~data_dir ~pvm_kind reveal_map
+      hash =
     let found_in_map =
       match reveal_map with
       | None -> None
@@ -45,7 +46,8 @@ module Make_fueled (F : Fuel.S) : FUELED_PVM with type fuel = F.t = struct
     in
     match found_in_map with
     | Some data -> return data
-    | None -> Reveals.get ~dac_client ~data_dir ~pvm_kind hash
+    | None ->
+        Reveals.get ~dac_client ~pre_images_endpoint ~data_dir ~pvm_kind hash
 
   type eval_completion =
     | Aborted of {state : pvm_state; fuel : fuel; current_tick : int64}
@@ -107,6 +109,7 @@ module Make_fueled (F : Fuel.S) : FUELED_PVM with type fuel = F.t = struct
           let*! data =
             get_reveal
               ~dac_client:node_ctxt.dac_client
+              ~pre_images_endpoint:node_ctxt.config.pre_images_endpoint
               ~data_dir:node_ctxt.data_dir
               ~pvm_kind:node_ctxt.kind
               reveal_map
@@ -228,6 +231,7 @@ module Make_fueled (F : Fuel.S) : FUELED_PVM with type fuel = F.t = struct
           let* data =
             get_reveal
               ~dac_client:node_ctxt.dac_client
+              ~pre_images_endpoint:node_ctxt.config.pre_images_endpoint
               ~data_dir:node_ctxt.data_dir
               ~pvm_kind:node_ctxt.kind
               reveal_map