+++ src/proto_018_Proxford/lib_sc_rollup_node/reveals.mli
@@ -54,9 +54,10 @@ type source =
   | File of string
       (** A file name whose associated file contains the whole data *)
 
-(** [get ~dac_client ~data_dir ~pvm_name ~hash] retrieves the data associated
-    with the reveal hash [hash] from disk. If the data is not already on disk,
-    it will be retrieved with the [dac_client]. If a May fail with:
+(** [get ~dac_client ~pre_images_endpoint ~data_dir ~pvm_name ~hash] retrieves
+    the data associated with the reveal hash [hash] from disk. If the data is
+    not already on disk, it will be retrieved with the [dac_client] or from an
+    HTTP service at [pre_images_endpoint]. May fail with:
     {ul
       {li [Wrong_hash {found; expected}] where [expected = hash], and
         [found <> hash], if the data is retrieved and hashes to the wrong
@@ -69,6 +70,7 @@ type source =
    } *)
 val get :
   dac_client:Dac_observer_client.t option ->
+  pre_images_endpoint:Uri.t option ->
   data_dir:string ->
   pvm_kind:Kind.t ->
   Protocol.Sc_rollup_reveal_hash.t ->